HPCC-12513 Clean up various windows warnings

+++ esp/tools/soapplus/http.cpp
@@ -1138,7 +1138,7 @@ int HttpClient::sendStressRequest(StringBuffer& request, HttpStat* stat)
     if(http_tracelevel >= 10)
         fprintf(m_logfile, "%s%s%s", sepstr, request.str(), sepstr);
 
-    __int64 start = msTick();
+    unsigned start = msTick();
 
     Owned<CSimpleSocket> sock = new CSimpleSocket(m_ssctx.get(), m_logfile);
     
@@ -1169,7 +1169,7 @@ int HttpClient::sendStressRequest(StringBuffer& request, HttpStat* stat)
         }
     }
     sock->close();
-    __int64 end = msTick();
+    unsigned end = msTick();
     int duration = end - start;
 
     if(http_tracelevel >= 5)
@@ -1461,7 +1461,7 @@ int HttpClient::sendRequest(StringBuffer& req, IFileIO* request_output, IFileIO*
     if(request_output)
         request_output->write(0, request.length(), request.str());
 
-    __int64 start1 = msTick();
+    unsigned start1 = msTick();
 
     SocketEndpoint ep;
     ep.set(m_host.str(), m_port);
@@ -1523,7 +1523,7 @@ int HttpClient::sendRequest(StringBuffer& req, IFileIO* request_output, IFileIO*
     socket->shutdown();
     socket->close();
 
-    __int64 end1 = msTick();
+    unsigned end1 = msTick();
 
     int duration = end1 - start1;
